Add Simplified CommonJS wrapper support to requirejs plugin.
In the requirejs plugin check to see if there is only one argument, and if it is a function expression with 3 args named require, exports, and module.  If so, treat it as a Simplified CommonJS module, and get the modules type from either the return value, or the exports argument.  Also create a Fn to represent the require function passed to the module so that type information is available for all the synchronous require calls in this style module.

The Simplified CommonJS wrapper is described at http://requirejs.org/docs/api.html#cjsmodule