Don't use (declare (special ...))
It doesn't work well with lexical-binding
(http://lists.gnu.org/archive/html/help-gnu-emacs/2014-07/msg00485.html),
and the code works just as well in every instance where it's been used
here.

#350

Note: the warning

tern.el:37:27:Warning: reference to free variable `url-http-process'

is displayed during byte-compilation, but it's been displayed before
this patch, as well. We can make it go away with a strategic `defvar'
as well, but the code works fine as it is anyway.