Adjust .jshintrc, try to get code to lint, give up
It complains about a number of things that are entirely valid,
or not even the case (see 'return after return' warning for lib/def.js).

It did catch a bug though, so there's an upside. Doesn't seem to be
configurable enough yet to suppress the bogus warnings, so I don't see
routine use working out yet.