Inject function type for tern.registerFunction handler
This PR inject function type to the tern.registerFunction handler. The function type can be usefull to retrieve argument type of the function.

Today we can do that with self, but we need hard coded the function name in the handler : 

var fnType = self.hasProp("MyFunction").getType();

with this PR, fnType can be retrieved directly in the handler with hard coding the function name.