Add from start/end of the node where find definition is executed
In Eclipse Tern IDE, when user wish to find definition, it uses Ctrl+Click on a node which call findDef (it is called Hyperlink). I need to underline this node and this PR gives me the capability to find the start/end of the node to underline.

This PR is simple, but after it should be cool to improve it to customize the from start/end like typeAt.