Allow absolute paths to be used in query
The primary use-case for this is that users of the API don't (or, shouldn't)
need to know anything about `.tern-project` and tern's internal project
structure. The simplest thing for editors to do is to just supply the abosolute
path to dirty buffers (in the files section of the query) and have tern
determine if/that they are also included by things like require.js.

A typical example might be a file containing `require('lib/test.js', ...)` and
an API user supplying an unsaved version of 'lib/test.js' with its absolute path
as the filename.