Make it work with both locally and globally installed plugins
The previous commit break the case where plugins are installed in
the  tern directory (i.e. going to the directory where tern is
installed and running npm install tern-plugin).
This commit supports both cases, where the plugin are installed
either in the tern directory or in the project root directory.